Move UE import funcs into imgspc framework
