+++ hooks/tk-multi-breakdown/unreal/scene_operations.py
@@ -113,27 +113,8 @@ class BreakdownSceneOperations(Hook):
         generated by the scan_scene hook above. The path key now holds
         the that each node should be updated *to* rather than the current path.
         """
-        engine = sgtk.platform.current_engine()
-
-        # tk-unreal from github releases
-        tk_unreal_actions = os.path.join(
-            engine.disk_location,
-            "hooks", "tk-multi-loader2", "tk-unreal_actions.py",
-        )
-
-        # Build the hook, with tk-unreal's hook as a parent
-        tk_imgspc_actions_hook = create_hook_instance(
-            [
-                tk_unreal_actions,
-                # resolve path to this config
-                engine._TankBundle__resolve_hook_expression(
-                    None,
-                    "{config}/tk-multi-loader2/unreal/scene_actions.py"
-                )[0]
-            ],
-            self,
-        )
-        tk_imgspc_actions_hook.parent.engine = engine
+        imgspc_fw = self.load_framework("tk-framework-imgspc")
+        imgspc_ue_imp = imgspc_fw.import_module("unreal.import_utils")
 
         for item in items:
 
@@ -165,7 +146,7 @@ class BreakdownSceneOperations(Hook):
 
             sg_publish_data = publishes[new_source_file_path]
 
-            tk_imgspc_actions_hook._import_to_content_browser(
+            imgspc_ue_imp.import_to_content_browser(
                 new_source_file_path,
                 sg_publish_data,
             )
